# [IMP] export: improve export memory footprint