@robodoo 2025-02-25
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/199218.png)](https://mergebot.odoo.com/odoo/odoo/pull/199218)